+++ examples/appendix_a/pong.rs
@@ -192,11 +192,27 @@ fn initialise_score(world: &mut World) {
         (),
         &world.read_resource(),
     );
-    let p1_transform =
-        UiTransform::new("P1".to_string(), Anchor::TopMiddle, -50., 50., 1., 55., 50.);
+    let p1_transform = UiTransform::new(
+        "P1".to_string(),
+        Anchor::TopMiddle,
+        Anchor::Middle,
+        -50.,
+        50.,
+        1.,
+        55.,
+        50.,
+    );
 
-    let p2_transform =
-        UiTransform::new("P2".to_string(), Anchor::TopMiddle, 50., 50., 1., 55., 50.);
+    let p2_transform = UiTransform::new(
+        "P2".to_string(),
+        Anchor::TopMiddle,
+        Anchor::Middle,
+        50.,
+        50.,
+        1.,
+        55.,
+        50.,
+    );
 
     let p1_score = world
         .create_entity()