+++ docs/CHANGELOG.md
@@ -78,6 +78,7 @@ extra bounds from `AnimatablePrefab` and `AnimationSetPrefab` ([#1435])
 * Added generic parameter type to `Transform` to configure floating point precision. ([#1334])
 * `NetConnection` is automatically created when client starts sends data to server. ([#1539])
 * User will receive `NetEvent::Connected` on new connection and `NetEvent::Dissconnected` on disconnect. ([#1539])
+* Added a `pivot` field to `UiTransform`. ([#1571])
 
 ### Removed
 - Removed all `NetEvent's` because they were not used. ([#1539])
@@ -146,6 +147,7 @@ extra bounds from `AnimatablePrefab` and `AnimationSetPrefab` ([#1435])
 [#1538]: https://github.com/amethyst/amethyst/pull/1538
 [#1539]: https://github.com/amethyst/amethyst/pull/1543
 [#1568]: https://github.com/amethyst/amethyst/pull/1568
+[#1571]: https://github.com/amethyst/amethyst/pull/1571
 ## [0.10.0] - 2018-12
 
 ### Added